Make the icon of the current engine appear at the left of floating panel of ibus
And also show the full engine name as a tooltip on that icon.

Before only a generic gear wheel ⚙️ was shown, the same for all
engines, so one could not distinguish which engine was selected in the
floating panel.

Make the icon of the current engine appear at the left of floating panel of ibus