Let IBusM17nEngine inherit from IBusEngineSimple to enable compose support
Resolves: https://github.com/ibus/ibus-m17n/issues/51

Inspired by the similar change in ibus-hangul:

https://github.com/libhangul/ibus-hangul/commit/afb508f0f858054b5172b87917ce5d873693b424

ibus-hangul has an English and a Hangul mode and enables compose only in English mode.

ibus-m17n has only one mode so it has to enable compose always.

So it is possible that there is already a non-empty preedit when a
compose sequence starts.  If that is the case, commit the current
content of the preedit and handle the compose sequence.