Fix option “☑️ Use US keyboard layout”
Resolves: https://github.com/ibus/ibus-m17n/issues/66

The change to the keyval of the US layout needs to be done **before**
doing the Compose support, i.e.  **before**:

   IBUS_ENGINE_CLASS (parent_class)->process_key_event (engine, keyval, keycode, modifiers)

If the keyval is not yet translated before the Compose support is
done, then the Compose support cannot work as intended.

On top of that, if a non-ASCII keyval is fed into the parent class
(IBusEngineSimple), then IBusEngineSimple usually commits it
immediately. That had the effect that when a Russian layout was used
together with the option

☑️ Use US keyboard layout

all the Cyrillic letters were committed immediately before they could
be translated into the respective key values on the US layout.