Do not hide empty preedit text before committing text
Resolves: https://github.com/ibus/ibus-m17n/pull/64

Together with ibus >= 1.5.29 this solves the preedit flicker
problem reported in https://github.com/ibus/ibus/issues/2536
perfectly.

Thank you very much to Takao Fujiwara for the pull request.

I rewrote it a bit to make it easier for me to understand.

Hiding preedit text before committing text causes a reset of the
cursor position.

minput_filter() sometimes causes two calls to
ibus_m17n_engine_update_preedit(), the first one with a non-empty
preedit and then a second one with an empty preedit if text to commit
was produced.  But the call to ibus_m17n_engine_commit_string ()
happens a bit later and updating the empty preedit before the commit
causes the cursor to jump left and then right again when the commit is
done (This cursor jump is not always reproducible, it depends on speed
and may be invisible).

This also reverts the b74c6a68453e5a8132c7f2c104c446c8e5604963
(the workaround for the flicker).