Prefer commit to "return FALSE;" for space, and other keys where the msymbol name is exactly one character, add a sleep for Return
Resolves: https://github.com/ibus/ibus-m17n/issues/72
(Reported on Fedora 39 Gnome Wayland)

See also this comment: https://github.com/ibus/ibus-m17n/issues/72#issuecomment-1828601107

Carlos says there that mutter currently assumes that there is never
more than only one commit on a single key event.  He suggests that I
could concatenate commit strings and commit once instead of using
multiple commits. For the problem reported in
https://github.com/ibus/ibus-m17n/issues/72 this workaround is
possible and might have no disadvantages. However, I still think that
more than one commit per key event should be allowed.

Doing only a single commit and then 'return TRUE;` avoids reordering in mutter.

That trick works well for space and other keys which have a Unicode
value, are not Control characters and when no modifiers are pressed
(When modifiers are pressed, there is no way to add that information
to the commit string).

The Return and KP_Enter key events are kind of a special case, even though
these are control characters, appending "\n" to the commit string works in
many cases. But not in all cases:

- older versions of firefox insert nothing when "\n" is committed.

- Messengers like Telegram, Zalo, ... which usually send a message
  when Return is typed don’t do that when the event is not passed
  through but "\n" is committed. Zalo inserts a space instead,
  Telegram does nothing.

- In some search entries, for example in the search entry in
  LibreOffice, passing the key event Return starts the search but
  committing "\n" does not.  In the search entry of Libreoffice,
  committing "\n" inserts something which looks like a ⏎ RETURN
  SYMBOL.

For the cases, where the trick to append to the commit string cannot
be used, a sleep of 0.1 seconds between the commit and the passing of
the key event to the application helps Mutter to guess which actions
belong together and get the order right. So this mostly "fixes" the
problem as well. A sleep is not so nice, but these other cases are
rare, so the sleep should not hurt much.